pyproject.toml
[project]
name = "byte-ai-cli"
version = "0.7.1"
description = "A human-in-the-loop AI coding agent that keeps you in control."
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python",
    "Topic :: Software Development",
]
dependencies = [
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.14.2",
    "click>=8.2.1",
    "gitpython>=3.1.45",
    "markdownify>=1.2.0",
    "pathspec>=0.12.1",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pydoll-python>=2.8.2",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "watchfiles>=1.1.0",
    "loguru>=0.7.3",
    "pygments>=2.19.2",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.1",
    "langchain-anthropic>=1.0.1",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.2",
    "langchain-core>=1.0.2",
    "lsp-client>=0.3.1",
]

[dependency-groups]
dev = [
    "grandalf>=0.8",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pytest-recording>=0.13.4",
    "python-semantic-release>=9.21.1",
    "ruff>=0.12.12",
    "ssort>=0.16.0",
    "tomli>=2.2.1",
    "ty>=0.0.4",
]

mkdocs = ["mkdocs>=1.6.1", "mkdocs-material>=9.6.21", "mkdocstrings>=0.30.1"]

[project.urls]
Homepage = "https://github.com/UseTheFork/byte"

[project.scripts]
byte = "byte.main:cli"

[tool.uv.build-backend]
module-name = "byte"

[tool.ty.environment]
python-version = "3.12"


[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["src/tests"]
markers = ["vcr: record/replay HTTP via VCR"]
norecursedirs = "tmp.* build benchmark _site OLD"
addopts = "--record-mode=once"
# addopts = -p no:warnings --show-capture=no
# testpaths =
#     tests/domain
# ; env =
# ;     BYTE_ANALYTICS=false

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    # keep-sorted start
    "E",
    "F",
    "FURB",
    "G",
    "I",
    "ICN",
    "ISC",
    "LOG",
    "PIE",
    "PT",
    "Q",
    "RSE",
    "RUF",
    "T10",
    "UP",
    # keep-sorted end
]
ignore = [
    # keep-sorted start
    "E111",
    "E114",
    "E117",
    "E501",
    "E712",
    "ISC001",
    "PIE790",
    "Q001",
    "Q002",
    "Q003",
    "RUF005",
    "RUF006",
    "RUF012",
    "UP006",
    "UP007",
    "UP035",
    "UP045",
    "W191",
    # keep-sorted end
]

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
combine-as-imports = true

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[tool.semantic_release]
version_toml = [
    "pyproject.toml:project.version",
    ".src/byte/__init__.py:__version__",
]
build_command = """
  uv lock --upgrade-package byte-ai-cli
  uv build
"""
commit_subject = "chore: release v{version}"
commit_message = "chore: release v{version}\n\n{changelog}"
commit_parser = "conventional"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = ["^Bump .*", "^Update .*", "^Merge pull request.*"]
mode = "init"
insertion_flag = "<!-- version list -->"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "style",
    "refactor",
    "ops",
    "test",
]
allowed_tags = [
    "feat",
    "fix",
    "perf",
    "build",
    "ops",
    "chore",
    "ci",
    "docs",
    "style",
    "refactor",
    "test",
]
default_bump_level = 0
parse_squash_commits = false
ignore_merge_commits = true

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
